requirements.txt
asgiref==3.9.1
Django==4.2.0
Pillow==10.0.0
whitenoise==6.5.0
sqlparse==0.5.3

# Telegram Bot
python-telegram-bot==21.0.1
